attempting to fix RBG conversion

+++ functions.py
@@ -249,8 +249,11 @@ def convert_grayscale_to_rgb(input_root, output_root):
                 # Replicate the single channel to create three identical channels
                 rgb_array = np.stack((grayscale_array,) * 3, axis=-1)
 
+                # Ensure the data type is uint8
+                rgb_array = rgb_array.astype('uint8')
+                
                 # Convert the NumPy array back to an image
-                rgb_image = Image.fromarray(rgb_array.astype('uint8'))
+                rgb_image = Image.fromarray(rgb_array)
 
                 # Save the RGB image to the output directory
                 output_image_path = os.path.join(output_label_dir, filename)
